+++ apps/docs/src/content/03-components/overlays/light-box/examples/imageTrigger.tsx
@@ -1,0 +1,23 @@
+import {
+  Button,
+  Image,
+  LightBox,
+  LightBoxTrigger,
+} from "@mittwald/flow-react-components";
+
+<LightBoxTrigger>
+  <Button>
+    <Image
+      width={100}
+      withBorder
+      alt="mittwald"
+      src="https://mittwald.github.io/flow/assets/mittwald_logo_rgb.jpg"
+    />
+  </Button>
+  <LightBox>
+    <Image
+      alt="mittwald"
+      src="https://mittwald.github.io/flow/assets/mittwald_logo_rgb.jpg"
+    />
+  </LightBox>
+</LightBoxTrigger>;