+++ packages/remote-react-components/src/components/Form.tsx
@@ -1,10 +1,9 @@
 "use client";
-import { createRemoteComponent } from "@remote-dom/react";
+import { createRemoteComponent } from "@mfalkenberg/remote-dom-react";
 import { RemoteFormElement } from "@mittwald/flow-remote-elements";
-export { type RemoteFormElement } from "@mittwald/flow-remote-elements";
 import type { ComponentProps, FC } from "react";
-import React from "react";
 import { startTransition } from "react";
+export { type RemoteFormElement } from "@mittwald/flow-remote-elements";
 
 const BaseForm = createRemoteComponent("flr-form", RemoteFormElement, {
   eventProps: {