+++ packages/remote-react-components/src/components/RemoteRoot.tsx
@@ -1,19 +1,22 @@
 "use client";
-import { type FC, Suspense } from "react";
-import React from "react";
-import { useIsMounted } from "@/hooks/useIsMounted";
 import type { RootClientProps } from "@/components/RemoteRootClient";
+import { useIsMounted } from "@/hooks/useIsMounted";
+import React, { type FC, Suspense } from "react";
 
 const ClientComponent = React.lazy(() => import("./RemoteRootClient"));
 
 export const RemoteRoot: FC<RootClientProps> = (props) => {
-  const { children, ...rest } = props;
   const isMounted = useIsMounted();
-  return isMounted ? (
-    <Suspense>
-      <ClientComponent {...rest}>{children}</ClientComponent>
+
+  if (!isMounted) {
+    return null;
+  }
+
+  return (
+    <Suspense fallback={props.fallback}>
+      <ClientComponent {...props} />
     </Suspense>
-  ) : null;
+  );
 };
 
 export default RemoteRoot;